+++ R/st_nn.R
@@ -1,21 +1,25 @@
 #' Nearest Neighbor Search for Simple Features
 #'
 #' The function returns the indices of layer \code{y} which are nearest neighbors of each feature of layer \code{x}. The number of nearest neighbors \code{k} and the search radius \code{maxdist} can be modified.\cr\cr
-#' The function has three modes of operation -
+#' The function has three modes of operation:
 #' \itemize{
-#' \item{lon-lat points - Calculation using C code from \code{GeographicLib}, similar to \code{sf::st_distance}}
-#' \item{projected points - Calculation using \code{RANN::nn2}, a fast search method based on the ANN C++ library}
-#' \item{lines or polygons, either lon-lat or projected - Calculation based on \code{sf::st_distance}}
+#' \item{lon-lat points—Calculation using C code from \code{GeographicLib}, similar to \code{sf::st_distance}}
+#' \item{projected points—Calculation using \code{RANN::nn2}, a fast search method based on the ANN C++ library}
+#' \item{lines or polygons, either lon-lat or projected—Calculation based on \code{sf::st_distance}}
 #' }
 #'
 #' @param x Object of class \code{sf} or \code{sfc}
 #' @param y Object of class \code{sf} or \code{sfc}
-#' @param sparse logical; should a sparse index list be returned (TRUE) or a dense logical matrix? See below.
-#' @param k The maximum number of nearest neighbors to compute. Default is \code{1}, meaning that only a single point (nearest neighbor) is returned
+#' @param sparse \code{logical}; should a sparse index list be returned (\code{TRUE}, the default) or a dense logical matrix? See below. This only affects the
+#' @param k The maximum number of nearest neighbors to compute. Default is \code{1}, meaning that only a single point (nearest neighbor) is returned.
 #' @param maxdist Search radius (in meters). Points farther than search radius are not considered. Default is \code{Inf} meaning that search is unconstrained
-#' @param returnDist logical; whether to return a matrix with the distances between detected neighbors
-#' @param progress Display progress bar? (default `TRUE`)
-#' @return If \code{sparse=FALSE}, returned object is a logical matrix with element \code{[i,j]} being \code{TRUE} when \code{y[j, ]} is a neighbor of \code{x[i]}; if \code{sparse=TRUE} (the default), a sparse list representation of the same matrix is returned, with list element \code{i} a numeric vector with the indices \code{j} of neighboring features from \code{y} for the feature \code{x[i, ]}, or \code{integer(0)} in case there are no neighbors. If \code{returnDists=TRUE} the function returns a \code{list}, with the first element as specified above, and the second element the matrix of distances (in meters) between each pair of detected neighbors.
+#' @param returnDist \code{logical}; whether to return a second \code{list} with the distances between detected neighbors.
+#' @param progress Display progress bar? (default \code{TRUE})
+#' @return  \itemize{
+#' \item{If \code{sparse=TRUE} (the default), a sparse \code{list} with list element \code{i} being a numeric vector with the indices \code{j} of neighboring features from \code{y} for the feature \code{x[i,]}, or \code{integer(0)} in case there are no neighbors.}
+#' \item{If \code{sparse=FALSE}, a \code{logical} matrix with element \code{[i,j]} being \code{TRUE} when \code{y[j,]} is a neighbor of \code{x[i]}.}
+#' \item{If \code{returnDists=TRUE} the function returns a \code{list}, with the first element as specified above, and the second element a sparse \code{list} with the distances (in meters) between each pair of detected neighbors corresponding to the sparse \code{list} of indices.}
+#' }
 #' @references C. F. F. Karney, GeographicLib, Version 1.49 (2017-mm-dd), \url{https://geographiclib.sourceforge.io/1.49}
 #' @export
 #'
@@ -27,6 +31,7 @@
 #'
 #' cities = st_transform(cities, 32636)
 #' towns = st_transform(towns, 32636)
+#' water = st_transform(water, 32636)
 #'
 #' # Nearest town
 #' st_nn(cities, towns)
@@ -56,7 +61,7 @@
 #' st_join(towns, cities, st_nn, k = 1)
 #'
 #' # Polygons to polygons
-#' st_nn(water, water, k = 4)
+#' st_nn(water, towns, k = 4)
 #'
 #' # Large example - Geo points
 #' n = 1000
@@ -84,16 +89,18 @@
 #' end - start
 #'
 #' # Large example - Polygons
-#' n = 1000
+#' set.seed(1)
+#' n = 50
 #' x = data.frame(
 #'   lon = (runif(n) * 2 - 1) * 70,
 #'   lat = (runif(n) * 2 - 1) * 70
 #' )
 #' x = st_as_sf(x, coords = c("lon", "lat"), crs = 4326)
 #' x = st_transform(x, 32630)
-#' x = st_buffer(x, 1000)
+#' x = st_buffer(x, 1000000)
 #' start = Sys.time()
 #' result = st_nn(x, x, k = 3)
+#' result[1:3]
 #' end = Sys.time()
 #' end - start
 #'
@@ -118,8 +125,6 @@ st_nn = function(x, y, sparse = TRUE, k = 1, maxdist = Inf, returnDist = FALSE, 
     stop("'x' and 'y' needs to be in the same CRS")
 
   # Determine geometry type and projection
-
-  # Check that 'x' and 'y' are 'POINT'
   if(!class(x)[1] == "sfc_POINT" | !class(y)[1] == "sfc_POINT") {
     result = .st_nn_poly(x, y, k, maxdist, progress)
   } else {
@@ -130,17 +135,13 @@ st_nn = function(x, y, sparse = TRUE, k = 1, maxdist = Inf, returnDist = FALSE, 
     }
   }
 
+  # Returned sparse lists
   ids = result[[1]]
-  dist_matrix = result[[2]]
+  dists = result[[2]]
 
-  # To sparse
-  ids = split(ids, 1:nrow(ids))
-  ids = lapply(ids, function(x) c(x[!is.na(x)]))
-  names(ids) = NULL
-  result = ids
-
-  # To dense matrix
+  # To dense matrix?
   if(!sparse) {
+    # ids
     m = matrix(
       FALSE,
       nrow = length(x),
@@ -150,8 +151,9 @@ st_nn = function(x, y, sparse = TRUE, k = 1, maxdist = Inf, returnDist = FALSE, 
     ids = m
   }
 
+  # Attach distances?
   if(returnDist) {
-    result = list(nn = ids, dist = dist_matrix)
+    result = list(ids, dists)
   } else {
     result = ids
   }