+++ R/st_nn_pnt_geo.R
@@ -8,41 +8,43 @@
   maxdist = as.numeric(maxdist)
 
   ids = matrix(NA, nrow = x_features, ncol = k)
-  dist_matrix = matrix(NA, nrow = x_features, ncol = k)
+  dists = matrix(NA, nrow = x_features, ncol = k)
 
   # Progress bar
-  if(progress) {
-    pb = utils::txtProgressBar(min = 0, max = x_features, initial = 0, style = 3)
-  }
+  if(progress) pb = utils::txtProgressBar(min = 0, max = x_features, initial = 0, style = 3)
 
   for(i in 1:x_features) {
 
-    dists = rep(NA, y_features)
+    dists1 = rep(NA, y_features)
 
     for(j in 1:y_features) {
-
-      dists[j] = .Call(addr_dist_one, c(x_coord[i, ], y_coord[j, ]))
-
+      dists1[j] = .Call(addr_dist_one, c(x_coord[i, ], y_coord[j, ]))
     }
 
-    ids1 = order(dists)[1:k]
+    ids1 = order(dists1)[1:k]
     ids[i, ] = ids1
-    dist_matrix[i, ] = dists[ids1]
+    dists[i, ] = dists1[ids1]
 
     # Progress
-    if(progress) {
-      utils::setTxtProgressBar(pb, i)
-    }
+    if(progress) utils::setTxtProgressBar(pb, i)
 
   }
 
-  ids[dist_matrix > maxdist] = NA
-  dist_matrix[is.na(ids)] = NA
+  ids[dists > maxdist] = NA
+  dists[is.na(ids)] = NA
 
-  if(progress) {
-    cat("\n")
-  }
-  return(list(ids, dist_matrix))
+  # From n*k 'matrix' to sparse 'list'
+  ids = split(ids, 1:nrow(ids))
+  ids = lapply(ids, function(x) c(x[!is.na(x)]))
+  names(ids) = NULL
+  dists = split(dists, 1:nrow(dists))
+  dists = lapply(dists, function(x) c(x[!is.na(x)]))
+  names(dists) = NULL
+
+  if(progress) cat("\n")
+
+  # Return sparse lists
+  return(list(ids, dists))
 
 }